chore: throwback double precision

+++ src/sd_finalProducts.cpp
@@ -36,7 +36,7 @@ void net_radiation_24h_function_worker(Products products, float Rs24h_val, float
     for (int r = start_row; r < end_row; ++r) {
         for (int c = 0; c < products.width_band; ++c) {
             int i = r * products.width_band + c;
-            products.net_radiation_24h[i] = (1.0f - products.albedo[i]) * Rs24h_val - FL_val * Rs24h_val / Ra24h_val;
+            products.net_radiation_24h[i] = (1.0 - products.albedo[i]) * Rs24h_val - FL_val * Rs24h_val / Ra24h_val;
         }
     }
 }
@@ -50,13 +50,13 @@ string net_radiation_24h_function(Products products, Station station, MTL mtl)
     initial_time = duration_cast<nanoseconds>(system_clock::now().time_since_epoch()).count();
     begin = system_clock::now();
 
-    float FL = 110.0f;
-    float dr = (1.0f / mtl.distance_earth_sun) * (1.0f / mtl.distance_earth_sun);
-    float sigma = 0.409f * sinf(((2.0f * PI / 365.0f) * mtl.julian_day) - 1.39f);
-    float phi = (PI / 180.0f) * station.latitude;
-    float omegas = acosf(-tanf(phi) * tanf(sigma));
-    float Ra24h = (((24.0f * 60.0f / PI) * GSC * dr) * (omegas * sinf(phi) * sinf(sigma) + cosf(phi) * cosf(sigma) * sinf(omegas))) * (1000000.0f / 86400.0f);
-    float Rs24h = station.INTERNALIZATION_FACTOR * sqrtf(station.v7_max - station.v7_min) * Ra24h;
+    float FL = 110.0;
+    float dr = (1.0 / mtl.distance_earth_sun) * (1.0 / mtl.distance_earth_sun);
+    float sigma = 0.409 * sin(((2.0 * PI / 365.0) * mtl.julian_day) - 1.39);
+    float phi = (PI / 180.0) * station.latitude;
+    float omegas = acos(-tan(phi) * tan(sigma));
+    float Ra24h = (((24.0 * 60.0 / PI) * GSC * dr) * (omegas * sin(phi) * sin(sigma) + cos(phi) * cos(sigma) * sin(omegas))) * (1000000.0 / 86400.0);
+    float Rs24h = station.INTERNALIZATION_FACTOR * sqrt(station.v7_max - station.v7_min) * Ra24h;
 
     parallel_for(products.height_band, [&](int start_row, int end_row) {
         net_radiation_24h_function_worker(products, Rs24h, Ra24h, FL, start_row, end_row);
@@ -73,8 +73,8 @@ void evapotranspiration_24h_function_worker(Products products, int start_row, in
     for (int r = start_row; r < end_row; ++r) {
         for (int c = 0; c < products.width_band; ++c) {
             int i = r * products.width_band + c;
-            float temperature_celcius = products.surface_temperature[i] - 273.15f;
-            products.evapotranspiration_24h[i] = (86400.0f / ((2.501f - 0.0236f * temperature_celcius) * powf(10.0f, 6.0f))) * (products.latent_heat_flux[i] / (products.net_radiation[i] - products.soil_heat[i])) * products.net_radiation_24h[i];
+            float temperature_celcius = products.surface_temperature[i] - 273.15;
+            products.evapotranspiration_24h[i] = (86400.0 / ((2.501 - 0.0236 * temperature_celcius) * pow(10.0, 6.0))) * (products.latent_heat_flux[i] / (products.net_radiation[i] - products.soil_heat[i])) * products.net_radiation_24h[i];
         }
     }
 }
